refactor: U1 抽离 detection/shell/headless 共享模块
从 mux.ts（~2110 行）中抽离三类共享基础设施代码：

- src/detection.ts: muxLog / MuxBackend / hasCommand / muxPreference /
  is*Available / getMuxBackend / muxSetupHint / AGENT_MUXY_PANE_ID
- src/shell.ts: isFishShell / exitStatusVar / shellEscape / tailLines /
  sleepSync / envPositiveInteger
- src/headless.ts: HeadlessProcess / headlessProcesses / createHeadless* /
  closeHeadless* / sendHeadlessEscape / readHeadlessScreen* /
  isHeadlessMode / cleanupHeadlessProcesses / drainHeadlessProcess

mux.ts 改为 import + re-export，index.ts 一行不用改。
公开 API 导出符号集合（81 个）与基线完全一致。
26/26 单测 + typecheck + 门禁全部通过。
